[ performance ] increase laziness in log annotations
This change led to a drastic boost in performance when generation
large, complex data types by means of `forAll`. Since fields in `Log`
where not lazy, `Annotation`s where processed on every iteration,
resulting in the parser from pretty-show being run. Since this
information is only needed in case of a test failure, this
was a horrible waste of resources.